Apply deltas to Blazor WebAssembly on app start
In Blazor WebAssembly, refreshing the browser causes the wasm app to be restarted and go out of sync with
where dotnet-watch thinks the app is at w.r.t what deltas have been applied. This PR introduces a way for
Blazor WASM to stash away these deltas and re-apply them when the app starts.

The implementation relies on the browser refresh middleware to stash and retrieve the deltas. Doing so has a few advantages
- we no longer have to concern ourselves with cross-origin requests
- the implementation can be much more easily shared with Visual Studio.